+++ Number Theory/ModMultiplication.cpp
@@ -1,5 +1,5 @@
-ll mod_multiplication(ll a,ll b){ // (x*y)%mod with intermediate states = no overflow
-    ll res=0;
+int mod_multiplication(int a,int b){ // (x*y)%mod with intermediate states = no overflow
+    int res=0;
     while(b){
         if(b&1) res=(res+a)%mod;
         a=(a+a)%mod;
