+++ Number Theory/checkprime.cpp
@@ -1,4 +1,4 @@
-bool prime(ll a) {
+bool prime(int a) {
     if (a <= 1) return 0;
     if (a == 2 || a == 3) return 1;
     if (a % 2 == 0 || a % 3 == 0) return 0;
