+++ Number Theory/fastexponentiation.cpp
@@ -1,5 +1,5 @@
-ll qexp(ll a, ll b, ll m) {
-    ll res = 1;
+int qexp(int a, int b, int m) {
+    int res = 1;
     while (b) {
         if (b & 1) res = res * a % m;
         a = a * a % m;
