+++ Number Theory/modinverse.cpp
@@ -1,17 +1,17 @@
-ll exteuclid(ll a, ll b, ll& x, ll& y) {
+int exteuclid(int a, int b, int& x, int& y) {
     if (b == 0) {
         x = 1;
         y = 0;
         return a;
     }
-    ll x1, y1;
-    ll d = exteuclid(b, a % b, x1, y1);
+    int x1, y1;
+    int d = exteuclid(b, a % b, x1, y1);
     x = y1;
     y = x1 - y1 * (a / b);
     return d;
 }
-ll modinv(ll a, ll m) {
-    ll x, y;
+int modinv(int a, int m) {
+    int x, y;
     exteuclid(a, m, x, y);
     return (m + x % m) % m;
 }
