+++ Utilities/BinaryStringtoDecimal.cpp
@@ -1,9 +1,9 @@
-ll binaryToDecimal(string n) {
+int binaryToDecimal(string n) {
     string num = n;
-    ll dec_value = 0;
-    ll base = 1;
-    ll len = num.length();
-    for (ll i = len - 1; i >= 0; i--) {
+    int dec_value = 0;
+    int base = 1;
+    int len = num.length();
+    for (int i = len - 1; i >= 0; i--) {
         if (num[i] == '1')
             dec_value += base;
         base = base * 2;
