+++ Utilities/pairsort.cpp
@@ -1,8 +1,7 @@
-bool ffDECssINC(const pair<ll,ll> &a, 
-              const pair<ll,ll> &b) 
+bool ffDECssINC(const pair<int,int> &a, 
+              const pair<int,int> &b) 
 {   
     if(a.first!=b.first) return (a.first>b.first); 
     return (a.second < b.second); 
 } 
 // > : decreasing sort < : increasing sort
-
